Refactor / move test helpers from src/test to src/main/app/model/__tests__/helpers
