+++ src/main/app/model/__tests__/AgentProxy.test.ts
@@ -1,4 +1,4 @@
-import { createTestAgentProxy } from "../../../../test/agent-proxy-test-helper";
+import { createTestAgentProxy } from "./helpers/agent-proxy-test-helper";
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { AgentDTO } from "../../../../domain";
 import { AgentProxy } from "../AgentProxy";
