+++ src/main/app/model/__tests__/PlanProxy.test.ts
@@ -1,4 +1,4 @@
-import { createTestPlanProxy } from "../../../../test/plan-proxy-test-helper";
+import { createTestPlanProxy } from "./helpers/plan-proxy-test-helper";
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { PlanDTO } from "../../../../domain";
 import { PlanProxy } from "../PlanProxy";
