+++ src/main/app/model/__tests__/helpers/plan-proxy-test-helper.ts
@@ -1,5 +1,5 @@
 import { createTestPrismaClient } from "./test-db-factory";
-import { PlanProxy } from "../main/app/model/PlanProxy";
+import { PlanProxy } from "../../PlanProxy";
 import { PrismaClient } from "db";
 
 /**
