+++ src/main/app/model/__tests__/helpers/project-proxy-test-helper.ts
@@ -1,5 +1,5 @@
 import { createTestPrismaClient } from "./test-db-factory";
-import { ProjectProxy } from "../main/app/model/ProjectProxy";
+import { ProjectProxy } from "../../ProjectProxy";
 import { PrismaClient } from "db";
 
 /**
