+++ src/main/app/model/__tests__/helpers/task-proxy-test-helper.ts
@@ -1,5 +1,5 @@
 import { createTestPrismaClient } from "./test-db-factory";
-import { TaskProxy } from "../main/app/model/TaskProxy";
+import { TaskProxy } from "../../TaskProxy";
 import { PrismaClient } from "db";
 
 /**
